Update Mode Hessian for a boss Object (with PSD check)

Compute the numerical Hessian of the objective at the current best point (in original input space), verify it is negative semi-definite (NSD), and store it back into the boss_result.

Usage

update_hessian(
  boss_result,
  approach = "local.poly",
  local.poly.args = list(eps = NULL, bw = NULL, kernel = "Epanech"),
  num.args = list(method = "Richardson", method.args = list()),
  tol = 1e-08,
  ...
)

Arguments

boss_result

A S3 boss object returned by BOSS_modal(), which must contain design_points$x_original and objective_function.

approach

Character string specifying the method used for estimating the Hessian at the BO mode. Must be one of local.poly, num.obj, num.GP.

local.poly.args

List of additional arguments for the local.poly method.

num.args

List of additional arguments for either num.obj or num.GP method.

tol

Numeric tolerance for eigenvalues (default 1e-8).

...

Further arguments passed to num.obj or num.GP which will be further passed to numDeriv::hessian.

Value

The input boss_result with two new components:

mode_hessian: symmetric Hessian matrix at the mode.

mode: the mode point in original input space.

Details

The function numerically computes the hessian at the mode obtained from a boss Object using three different strategies:

1. Default: Locally weighted polynomial regression (local.poly): given a specified accuracy level local.poly.args$eps. The function then first draws a circle with radius local.poly.args$eps/2 around the mode from the boss object and check how many boss design points are within the circle. If the number is below \(n = (D+2)(D+1)\), additional uniformly distributed design points are added within the circle to reach \(n\) and evaluate at these additional points with boss_result$objective_function. Combining all design points in the circle, estimate the Hessian at boss mode by a locally weighted polynomial regression using estimate_hessian(). The default kernel bw argument for estimate_hessian() is set to local.poly.args$eps / 2*D^0.4. If local.poly.args$eps = NULL, it is obtained by $$\epsilon = \frac{1}{4\sqrt{\text{Tr}(-H_{GP})}},$$ where \(H_{GP}\) is the hessian at the boss mode computed from numDeriv::hessian() on the GP surrogate boss_result$surrogate.

2. Brute-force numerical hessian (num.obj): directly estimate the hessian at the boss mode via numDeriv::hessian using the boss_result$objective_function.

3. Numerical hessian based on GP surrogate (num.GP): estimate the hessian at the boss mode via numDeriv::hessian using the boss_result$surrogate.

Note that local.poly balances between theoretical accuracy and computational budget. num.obj is the most computationally intense while num.GP is the cheapest, but does not have theoretical guarantee.