Approximately fill-in by supplying a Sobol-sequence in the essential support and filter against existing design points.

1. Based on the required fill-in distance h, compute the required number of quasi-uniform design points by \(n = {n_sample_max, 2*Vol(Original box)/h^D}\). 2. Generate a Sobol-sequence within the original box. 3. Truncate the Sobol-sequence within the ellipsoid. 4. Compute the distance from existing design points to the Sobol-sequence and remove Sobol-candidates that are within h of existing points. 5. Add in a maximum of max_add number selected fill-in Sobol-candidates to ensure optimal covering over the ellipsoid. 6. Recompute new fill-in based on the selected Sobol and existing design points.

Usage

fill_in(boss_result, h, max_add = 100, n_sample_max = 10000, verbose = 0)

Arguments

boss_result

A boss object with non-NULL essential_support and essential_design_points\$x_original, and containing lower and upper.

h

Numeric; target fill-in distance (>= 0).

max_add

Integer; maximum number of points to add (default 100).

n_sample_max

Integer; maximum number of Sobol-sequence candidates to be added (default 10000).

verbose

Integer in {0,1}; 0=silent, 1=summary.

Value

The updated boss_result, with its essential_design_points and fill_in refreshed.

Details

If the candidate pool is over max_add, or if fill_in cannot be reduced below h, a warning is issued.